+++ experiment/client-go/pkg/gowandb/run.go
@@ -9,8 +9,8 @@ import (
 	"github.com/segmentio/encoding/json"
 
 	"github.com/wandb/wandb/core/pkg/gowandb/opts/runopts"
-	"github.com/wandb/wandb/core/pkg/gowandb/runconfig"
 	"github.com/wandb/wandb/core/pkg/service"
+	"github.com/wandb/wandb/gowandb/runconfig"
 )
 
 type Settings map[string]interface{}
